requirements.txt
openai
pinecone-client
pandas
python-dotenv
tqdm
jupyter
notebook
scikit-learn
datasets
chromadb
sentence-transformers
pypdf
fpdf
streamlit